# move cursor on automated timer