Applications using the circuit breaker register traffic events with the Record() method. In the Record() method the existing Go circuit breaker calculates:
- error rate
- cursor expiry given current system time (and moves cursor if needed)

These calculations take a significant amount of time relative to the rest of the circuit breaker logic. This impacts the overall performance because the Record() method will be the most frequently used method on the circuit breaker. For high traffic applications the vast majority of Record() events will not result in moving the cursor, which means the cursor expiry calculations are redundant.

This PR suggests to defer this cursor expiry/movement logic and schedule it to only run exactly when it is needed using timers. This means the Record() method becomes much lighter and improves the overall circuit breaker performance.

### Benchmarks comparison

#### Record() benchmarks

Current (calculating if cursor needs to move in Record())
```
BenchmarkCircuitBreakerRecord-12        12323922                87.57 ns/op            0 B/op          0 allocs/op
```

New (Record() only registers success of failure events and half open state)
```
BenchmarkCircuitBreakerRecord-12        44685660                24.82 ns/op            0 B/op          0 allocs/op
```

#### End-to-end benchmarks

Current (calculating if cursor needs to move in Record())
```
BenchmarkCircuitBreaker-12      360152193              102.1 ns/op             0 B/op          0 allocs/op
```

New (deferring and calculating cursor movement on timer)
```
BenchmarkCircuitBreaker-12      880799961               40.33 ns/op            0 B/op          0 allocs/op
```

Overall these results indicate:
- ~70% improvement in Record() performance
- ~60% improvement in overall circuit breaker performance